+++ src/app/join/components/JoinHeader.client.tsx
