+++ src/app/join/funnels/step1/Step1Component.client.tsx
@@ -1,6 +1,8 @@
+'use client';
+
 import FormSection from './components/FormSection.client';
 import NoticeSection from './components/NoticeSection.client';
-import JoinTitleTextMessage from '../../components/JoinTitleTextMessage.server';
+import JoinTitleTextMessage from '../../components/JoinTitleTextMessage.client';
 import { Spacing } from '@/components/common/Spacing';
 
 export default function Step1Component() {
