+++ src/app/join/funnels/step1/components/NoticeSection.client.tsx
@@ -1,3 +1,4 @@
+'use client';
 import { Spacing } from '@/components/common/Spacing';
 import { TextList } from '@/components/TextList';
