+++ src/app/join/funnels/step2/Step2Component.client.tsx
@@ -1,6 +1,8 @@
+'use client';
+
 import InputForm from './components/InputForm.client';
 import SubmitSection from './components/SubmitSection.client';
-import JoinTitleTextMessage from '../../components/JoinTitleTextMessage.server';
+import JoinTitleTextMessage from '../../components/JoinTitleTextMessage.client';
 
 export default function Step2Component() {
   return (
