+++ src/app/join/funnels/step2/components/InputForm.client.tsx
@@ -1,7 +1,9 @@
+'use client';
+
 import AgreeSection from './AgreeForm.client';
 import SchoolSection from './SchoolSection.client';
 import { useFunnelContext } from '../../JoinFunnel';
-import { useJoinContext } from '@/app/join/components/JoinContext';
+import { useJoinContext } from '@/app/join/components/JoinContext.client';
 import { SignUpState } from '@/app/join/type';
 
 export default function InputForm() {
