+++ src/app/join/funnels/step2/components/SearchResultSection.client.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import type { SearchResultType } from '../type';
 
 interface SearchResultSectionProps {
