+++ src/app/join/funnels/step2/components/SubmitSection.client.tsx
@@ -1,5 +1,5 @@
 'use client';
-import { useJoinContext } from '../../../components/JoinContext';
+import { useJoinContext } from '../../../components/JoinContext.client';
 import { useFunnelContext } from '../../JoinFunnel';
 import { BottomFixedButton } from '@/components/common/Button';
