+++ src/app/join/funnels/step3/Step3Component.client.tsx
@@ -1,7 +1,9 @@
+'use client';
+
 import FormSection from './components/FormSection.client';
-import NoticeSection from './components/NoticeSection';
-import TimerContext from './components/TimerContext';
-import JoinTitleTextMessage from '../../components/JoinTitleTextMessage.server';
+import NoticeSection from './components/NoticeSection.client';
+import TimerContext from './components/TimerContext.client';
+import JoinTitleTextMessage from '../../components/JoinTitleTextMessage.client';
 import ModalContextProvider from '@/components/Modal/ModalContext.client';
 
 export default function Step3Component() {
