+++ src/app/join/funnels/step3/components/FormSection.client.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import CertificationForm from './CertificationForm.client';
 import EmailForm from './EmailForm.client';
