+++ src/app/join/funnels/step3/components/NoticeSection.client.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { Spacing } from '@/components/common/Spacing';
 import { TextList } from '@/components/TextList';
 import { INSTAGRAM_URL } from '@/constants';
