+++ src/app/join/funnels/step3/components/TimerContext.client.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useTimer } from '@/hooks/useTimer';
 import { createContext, useContext } from 'react';
