+++ src/app/join/funnels/step4/Step4Component.client.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import InputForm from './components/InputForm.client';
 
 export default function Step4Component() {
