+++ src/app/join/funnels/step5/Step5Component.client.tsx
@@ -1,6 +1,8 @@
+'use client';
+
 import InputForm from './components/InputForm.client';
-import JoinContentTextMessage from '../../components/JoinContentTextMessage.server';
-import JoinTitleTextMessage from '../../components/JoinTitleTextMessage.server';
+import JoinContentTextMessage from '../../components/JoinContentTextMessage.client';
+import JoinTitleTextMessage from '../../components/JoinTitleTextMessage.client';
 import { Spacing } from '@/components/common/Spacing';
 
 export default function Step5Component() {
