+++ src/app/join/funnels/step5/components/PersonalitySection.client.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import PersonalityItemServer from './PersonalityItem.client';
 import { personalityList } from '@/constants/personalityList';
